Fix polling
